+++ packages/deepclaw-agent/src/base-file-initializer.ts
@@ -1,10 +1,9 @@
 import { FileUtils } from '@deepclaw/node-utils';
-import { DEEPCLAW_MD, SKILLS, AGENTS_DIR } from './agent/paths';
+import { SKILLS, AGENTS_DIR } from './agent/paths';
 import { fileURLToPath } from 'url'
 import { dirname } from 'path'
 
 export function ensureBaseFiles() {
     const __dirname = dirname(fileURLToPath(import.meta.url))
-    FileUtils.copyResource(__dirname, DEEPCLAW_MD);
     FileUtils.copyResource(__dirname, SKILLS, AGENTS_DIR);
 }